+++ library/LocalRegions/SegExp.cpp
@@ -745,7 +745,7 @@ cout<<"deps/dx ="<<inarray_d0[i]<<"  deps/dy="<<inarray_d1[i]<<endl;
             const NekDouble           &inarray,
              Array<OneD, NekDouble>   &outarray)
         {
-            int nquad = m_base[0]->GetNumPoints();
+            size_t nquad = m_base[0]->GetNumPoints();
             
             if (m_base[0]->GetPointsType() != LibUtilities::eGaussGaussLegendre)
             {
@@ -957,8 +957,8 @@ cout<<"deps/dx ="<<inarray_d0[i]<<"  deps/dy="<<inarray_d1[i]<<endl;
             }
 
 
-            int nqb = nqe;
-            int nbnd= vertex;
+            size_t nqb = nqe;
+            size_t nbnd= vertex;
             m_elmtBndNormDirElmtLen[nbnd] = 
                     Array<OneD, NekDouble> {nqb, 0.0};
             Array<OneD, NekDouble>  &length =
